refactor: cleanup and organse cli commands
