+++ generator/extractor.go
